# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación realizado con Swagger, servicios funcionando con Laravel 8",
        "description": "Proyecto para la comunicación REST del sistema de Inventarios",
        "contact": {
            "name": "Correo de contacto",
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/SISHRS/backend/public"
        }
    ],
    "paths": {
...
        "/service/rest/v1/catalogo/status": {
            "get": {
                "tags": [
                    "Status"
                ],
                "summary": "Muestra todos los estados registrados en la base de datos.",
                "description": "Retorna un array de estados.",
                "operationId": "getStatus",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/catalogo/typebed": {
            "get": {
                "tags": [
                    "Type Bed"
                ],
                "summary": "Muestra todos los tipos de cama registradas en la base de datos.",
                "description": "Retorna un array de tipos de cama.",
                "operationId": "getTypeBeds",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Type Bed"
                ],
                "summary": "Crear un nuevo tipo de cama en el sistema.",
                "description": "Retorna el objeto del tipo de cama creado.",
                "operationId": "postTypeBed",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/catalogo/typebed/{typebed}": {
            "put": {
                "tags": [
                    "Type Bed"
                ],
                "summary": "Actualizar el tipo de cama seleccionado.",
                "description": "Retorna el objeto del tipo de cama actualizado.",
                "operationId": "updateTypeBed",
                "parameters": [
                    {
                        "name": "typebed",
                        "in": "path",
                        "description": "ID del tipo de cama para actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/catalogo/typebed/{typebed}   ": {
            "delete": {
                "tags": [
                    "Type Bed"
                ],
                "summary": "Eliminar el tipo de cama seleccionado.",
                "description": "Retorna el objeto del tipo de cama eliminado.",
                "operationId": "deleteTypeBed",
                "parameters": [
                    {
                        "name": "typebed",
                        "in": "path",
                        "description": "ID del tipo de cama para eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}